requirements.txt
apache-airflow===2.10.0
requests==2.32.3
duckdb==1.1.1
airflow-provider-duckdb== 0.2.0